add plpython3u for database; write a guide for huggingface new space and new model
